requirements.txt
httpx==0.27.2
openai==1.44.1
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1